chore: Rename existing Welcome to Dashboard
Our initial welcome screen has evolved into the Dashboard page. Now that we are talking about having a new welcome/initial setup page I'd like to rename the existing page to keep things clear.
