+++ packages/renderer/src/lib/dashboard/DashboardPage.svelte
