+++ packages/renderer/src/lib/dashboard/PreflightChecks.svelte
