+++ packages/renderer/src/lib/dashboard/ProviderConfigured.svelte
