+++ packages/renderer/src/lib/dashboard/ProviderDetectionChecksButton.svelte
