+++ packages/renderer/src/lib/dashboard/ProviderInstallationButton.svelte
