+++ packages/renderer/src/lib/dashboard/ProviderInstalled.svelte
