+++ packages/renderer/src/lib/dashboard/ProviderLogo.svelte
