+++ packages/renderer/src/lib/dashboard/ProviderNotInstalled.svelte
